feat: add endpoint for exchanging idporten tokens
- A new `/token` endpoint exchanges idporten tokens with self-issued
tokens that are signed  with a private, symmetric key. The token is only
meant to be used for requests to this very same application.
- The `/auth` and `/summary` endpoints will now accept requests
containing the aforementioned token. The `AuthTokenService` will
attempt to fetch both the `Oidc` (Loginservice) token, as well as the
`SelfIssued` token from a given request, though they will both be parsed
to the same `OidcTokenInfo` data class.
- In the event that both of the tokens are present in a request, the
most recently issued token will take precedence.

Co-Authored-By: Youssef Bel Mekki <youssef.bel.mekki@nav.no>