refactor: reduce size for self-issued tokens
These tokens are meant to be stored in cookies on the entire domain.
Too many large cookies may result in issues with webservers rejecting
requests, thus we should be conservative about the amount of data
that the token contains.

This commit removes some unused claims such as `jti` and `idp`,
which is generally used to prevent token replay and tracability
for the original token issuer, respectively:

- The `/auth` endpoint does not mandate any restrictions for
token replay, as this would require new tokens to be
exchanged/acquired for every request, which doesn't make sense
for this kind of service.
- The `/token` endpoint only accepts subject tokens from an already
defined IDP, thus it doesn't really make sense to include this in the
self-isused token either as it is implicitly known.